apps/pkm-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pkm-agent"
version = "0.1.0"
description = "AI-enhanced Personal Knowledge Management agent with TUI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "PKM Agent Team" }
]
keywords = ["pkm", "ai", "agent", "tui", "knowledge-management", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "textual>=0.47.0",
    "rich>=13.7.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.0",
    "openai>=1.10.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "watchdog>=4.0.0",
    "websockets>=12.0",
    "aiosqlite>=0.19.0",
    "python-frontmatter>=1.1.0",
    "tiktoken>=0.5.0",
    "tenacity>=8.2.0",
    "toml>=0.10.2",
    "rapidfuzz>=3.6.0",
    "aiohttp>=3.9.0",
    "psutil>=5.9.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]
ollama = [
    "ollama>=0.1.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
pkm-agent = "pkm_agent.cli:main"
pkma = "pkm_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pkm_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
